+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## XX.XX.XX
+## 24.7.6
 * Added support for localization of content blocks.
 
 * Mitigated an issue where visibility could have been wrongly assigned if a view was closed while going to background. (Experimental!)